# Conflicts:
#	api/src/Box9.Leds.Manager.DataAccess/Box9.Leds.Manager.DataAccess.csproj
#	api/src/Box9.Leds.Manager.Services/VideoPlayback/VideoPlaybackService.cs

+++ api/src/Box9.Leds.Manager.Services/AudioPlayback/AudioTrack.cs
@@ -27,7 +27,16 @@ namespace Box9.Leds.Manager.Services.AudioPlayback
 
         public void Dispose()
         {
-            File.Delete(AudioFilePath);
+            if (File.Exists(AudioFilePath))
+            {
+                try
+                {
+                    File.Delete(AudioFilePath);
+                }
+                catch
+                {
+                }               
+            }
         }
     }
 }
